feat     - Add calculation of tax (impot sur le revenu)     - Tuning of rendering of an invoice     - preparation of persitence via JPA

+++ src/main/java/com/softnaptics/training/domain/billing/utils/Period.java
@@ -26,7 +26,7 @@ public class Period {
         this.count = count;
     }
 
-    public static Period forDays(int count) {
+    public static Period forDays(double count) {
         return new Period(Type.DAY, count);
     }
